added comand name to packages
