+++ docs/modules/arc42-docs/pages/04-solution-strategy.adoc
@@ -1,5 +1,8 @@
 = 04 - Solution Strategy
 
+== Code Styleguide
+The source code of this application follows the link:https://go.dev/doc/effective_go[Effective Go] guide from the Go Website.
+
 // .Contents
 // A short summary and explanation of the fundamental decisions and solution strategies, that shape system architecture. It includes
